Move `finish_fn` into a dedicated module
