+++ bon-macros/src/builder/builder_gen/mod.rs
@@ -1,5 +1,6 @@
 mod builder_derives;
 mod builder_params;
+mod finish_fn;
 mod member;
 mod models;
 mod setters;
@@ -9,9 +10,7 @@ pub(crate) mod input_fn;
 pub(crate) mod input_struct;
 
 use crate::util::prelude::*;
-use member::{
-    Member, MemberOrigin, NamedMember, PositionalFnArgMember, RawMember, StartFnArgMember,
-};
+use member::{Member, MemberOrigin, NamedMember, RawMember, StartFnArgMember};
 use models::{
     AssocMethodCtx, AssocMethodReceiverCtx, BuilderGenCtx, FinishFn, FinishFnBody, Generics,
 };
@@ -459,146 +458,6 @@ impl BuilderGenCtx {
                         ::core::option::Option<#named_members_types>,
                     )*
                 )
-            }
-        }
-    }
-
-    fn finish_fn_member_expr(member: &Member) -> TokenStream {
-        let member = match member {
-            Member::Named(member) => member,
-            Member::Skipped(member) => {
-                return member
-                    .value
-                    .as_ref()
-                    .as_ref()
-                    .map(|value| quote! { (|| #value)() })
-                    .unwrap_or_else(|| quote! { ::core::default::Default::default() });
-            }
-            Member::StartFnArg(member) => {
-                let index = &member.index;
-                return quote! { self.__private_start_fn_args.#index };
-            }
-            Member::FinishFnArg(member) => {
-                return member.maybe_into_ident_expr();
-            }
-        };
-
-        let index = &member.index;
-
-        let member_field = quote! {
-            self.__private_named_members.#index
-        };
-
-        let param_default = member
-            .params
-            .default
-            .as_ref()
-            .map(|default| default.value.as_ref());
-
-        match param_default {
-            Some(Some(default)) => {
-                let has_into = member.params.into.is_present();
-                let default = if has_into {
-                    quote! { ::core::convert::Into::into((|| #default)()) }
-                } else {
-                    quote! { #default }
-                };
-
-                quote! {
-                    ::core::option::Option::unwrap_or_else(#member_field, || #default)
-                }
-            }
-            Some(None) => {
-                quote! {
-                    ::core::option::Option::unwrap_or_default(#member_field)
-                }
-            }
-            None => {
-                // For `Option` the default value is always `None`. So we can just return
-                // the value of the member field itself (which is already an `Option<T>`).
-
-                if member.is_special_option_ty() {
-                    return member_field;
-                }
-
-                quote! {
-                    unsafe {
-                        // SAFETY: we know that the member is set because we are in
-                        // the `finish` function because this method uses the trait
-                        // bounds of `IsSet` for every required member. It's also
-                        // not possible to intervene with the builder's state from
-                        // the outside because all members of the builder are considered
-                        // private (we even generate random names for them to make it
-                        // impossible to access them from the outside in the same module).
-                        ::core::option::Option::unwrap_unchecked(#member_field)
-                    }
-                }
-            }
-        }
-    }
-
-    fn finish_fn(&self) -> TokenStream {
-        let members_vars_decls = self.members.iter().map(|member| {
-            let expr = Self::finish_fn_member_expr(member);
-            let var_ident = member.orig_ident();
-
-            // The type hint is necessary in some cases to assist the compiler
-            // in type inference.
-            //
-            // For example, if the expression is passed to a function that accepts
-            // an impl Trait such as `impl Default`, and the expression itself looks
-            // like `Default::default()`. In this case nothing hints to the compiler
-            // the resulting type of the expression, so we add a type hint via an
-            // intermediate variable here.
-            let ty = member.norm_ty();
-
-            quote! {
-                let #var_ident: #ty = #expr;
-            }
-        });
-
-        let state_mod = &self.state_mod.ident;
-
-        let finish_fn_params = self
-            .members
-            .iter()
-            .filter_map(Member::as_finish_fn_arg)
-            .map(PositionalFnArgMember::fn_input_param);
-
-        let body = &self.finish_fn.body.generate(&self.members);
-        let asyncness = &self.finish_fn.asyncness;
-        let unsafety = &self.finish_fn.unsafety;
-        let must_use = &self.finish_fn.must_use;
-        let attrs = &self.finish_fn.attrs;
-        let finish_fn_vis = self
-            .finish_fn
-            .vis
-            .as_ref()
-            .unwrap_or(&self.builder_type.vis);
-        let finish_fn_ident = &self.finish_fn.ident;
-        let output = &self.finish_fn.output;
-
-        quote! {
-            #(#attrs)*
-            #[inline(always)]
-            #[allow(
-                // This is intentional. We want the builder syntax to compile away
-                clippy::inline_always,
-
-                // This lint flags any function that returns a possibly `!Send` future.
-                // However, it doesn't apply in the generic context where the future is
-                // `Send` if the generic parameters are `Send` as well, so we just suppress
-                // this lint. See the issue: https://github.com/rust-lang/rust-clippy/issues/6947
-                clippy::future_not_send,
-            )]
-            #must_use
-            #finish_fn_vis #asyncness #unsafety fn #finish_fn_ident(self, #(#finish_fn_params,)*) #output
-            where
-                // #( #where_bounds, )*
-                BuilderState: #state_mod::IsComplete
-            {
-                #(#members_vars_decls)*
-                #body
             }
         }
     }
